+++ tests/unit/hooks.test.ts
@@ -24,3 +24,10 @@ test('call a function after browser launched', done => {
     })
     .end(done)
 })
+
+test('call a function before browser exit', async () => {
+  const fn = jest.fn()
+  const instance = new Rize({ beforeExit: fn })
+  await instance.end()
+  expect(fn).toBeCalled()
+})